refactor(mobile): move input with label widget in the widget folder
