+++ mobile/lib/app/features/automation/view/singleton/settings.view.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:triggo/app/features/automation/bloc/automation/automation_bloc.dart';
 import 'package:triggo/app/features/automation/models/input.model.dart';
 import 'package:triggo/app/features/automation/view/singleton/input.view.dart';
-import 'package:triggo/app/routes/custom.router.dart';
+import 'package:triggo/app/features/automation/widgets/input_parameter.dart';
 import 'package:triggo/app/routes/routes_names.dart';
 import 'package:triggo/app/widgets/scaffold.triggo.dart';
 
@@ -22,7 +21,7 @@ class AutomationSettingsView extends StatelessWidget {
             padding: const EdgeInsets.all(4.0),
             child: ListView(
               children: [
-                AutomationLabelParameterWidget(
+                AutomationInputParameterWithLabel(
                   title: 'Label',
                   previewData: state.cleanedAutomation.label.isNotEmpty
                       ? state.cleanedAutomation.label
@@ -43,7 +42,7 @@ class AutomationSettingsView extends StatelessWidget {
                   ),
                 ),
                 SizedBox(height: 12.0),
-                AutomationLabelParameterWidget(
+                AutomationInputParameterWithLabel(
                   title: 'Description',
                   previewData: state.cleanedAutomation.description.isNotEmpty
                       ? state.cleanedAutomation.description
@@ -67,74 +66,6 @@ class AutomationSettingsView extends StatelessWidget {
           ),
         );
       },
-    );
-  }
-}
-
-class AutomationLabelParameterWidget extends StatelessWidget {
-  final String title;
-  final String? previewData;
-  final Widget input;
-  final bool disabled;
-
-  const AutomationLabelParameterWidget({
-    super.key,
-    required this.title,
-    required this.previewData,
-    required this.input,
-    this.disabled = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      behavior: HitTestBehavior.opaque,
-      onTap: () {
-        if (disabled) return;
-        Navigator.push(context, customScreenBuilder(input));
-      },
-      child: Container(
-        padding: EdgeInsets.all(12.0),
-        decoration: BoxDecoration(
-          color: disabled ? Colors.grey.shade300 : Colors.white,
-          borderRadius: BorderRadius.circular(8.0),
-        ),
-        child: Row(
-          children: [
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    title,
-                    style: Theme.of(context).textTheme.labelLarge!.copyWith(
-                          fontWeight: FontWeight.bold,
-                        ),
-                  ),
-                  Text(
-                    previewData == null ? 'No $title' : previewData!,
-                    style: Theme.of(context).textTheme.labelMedium,
-                  ),
-                ],
-              ),
-            ),
-            if (previewData == null)
-              Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 5.0),
-                child: Icon(
-                  Icons.warning_amber_rounded,
-                  color: Theme.of(context).colorScheme.onError,
-                  size: 30,
-                ),
-              ),
-            SvgPicture.asset(
-              "assets/icons/chevron-right.svg",
-              width: 24.0,
-              height: 24.0,
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
